Cast
git-svn-id: http://thebes.codingwell.net/repos/BC@287 23f83f56-a308-4d86-8998-a9fc3f2562b2

+++ Trunk/src/net/sf/backcomp/plugins/utils/CpuLimiter.java
@@ -138,7 +138,8 @@ public class CpuLimiter
 			final long cpuInterval = currentCpu - mLastCpuTime;
 			
 			final long targetEndTime =
-				( cpuInterval * HUNDRED_PERCENT ) / mCpuGoal + mLastCheck;
+				( cpuInterval * (long) HUNDRED_PERCENT ) / (long) mCpuGoal
+				+ mLastCheck;
 			
 			final long sleeptime =
 				( targetEndTime - currentTime ) / NANOS_IN_MILLI - 1L;
